Outer Border Started
Repeat the step of creating numOfCells and set them to grey. Then i can add numbers in. This way is easier as when it comes to the actionlistener part i can sort them as an array so whenever user clicks on the tile that has a number it wont change it will stay the same.